+++ ui/components/atoms/A009-AppBar/stories.js
@@ -6,7 +6,7 @@ import AppBar from '.'
 
 import notes from './readme.md'
 
-storiesOf('Atoms|A004 - AppBar', module)
+storiesOf('Atoms|A009 - AppBar', module)
   .addDecorator(withNotes)
   .add('App bar', () => <AppBar position='fixed' />, {
     notes: { markdown: notes }
