+++ ui/components/organisms/O001-Layout/index.js
@@ -4,7 +4,7 @@ import createPropTypes from 'json-schema-prop-types'
 import schema from './schema'
 
 import Appbar from '../../molecules/M001-AppBar'
-import ResponsiveDrawer from '../../molecules/M003-ResponsiveDrawer'
+import ResponsiveDrawer from '../../molecules/M002-ResponsiveDrawer'
 
 const O001 = props => {
   const { open, handleDrawerToggle, children } = props
